+++ itextsharp.layout/itextsharp/layout/splitting/DefaultSplitCharacters.cs
@@ -59,7 +59,7 @@ namespace iTextSharp.Layout.Splitting
 			{
 				return false;
 			}
-			int charCode = text.Get(glyphPos).GetUnicode();
+			int charCode = (int)text.Get(glyphPos).GetUnicode();
 			return (charCode <= ' ' || charCode == '-' || charCode == '\u2010' || (charCode >=
 				 0x2002 && charCode <= 0x200b) || (charCode >= 0x2e80 && charCode < 0xd7a0) || (
 				charCode >= 0xf900 && charCode < 0xfb00) || (charCode >= 0xfe30 && charCode < 0xfe50
